# swagger.yaml
openapi: 3.0.3
info:
  title: Redis Streams POC API
  description: |
    A comprehensive Node.js application demonstrating Redis Streams for real-time data processing with Express, PostgreSQL, and Sequelize.
    
    ## Features
    - 🚀 Redis Streams Integration with producer/consumer patterns
    - 📊 Consumer Groups for scalable and fault-tolerant message processing
    - 🗄️ PostgreSQL + Sequelize for persistent data storage
    - 🌐 RESTful API endpoints for triggering events
    - 📝 Complete audit trail of all events
    - 🔄 Real-time processing with event-driven architecture
    
    ## Redis Streams
    This API demonstrates various Redis Streams operations:
    - **XADD** - Publishing events to streams
    - **XREAD** - Simple event consumption
    - **XREADGROUP** - Consumer group processing
    - **XACK** - Message acknowledgment
    - **XGROUP CREATE** - Consumer group management
  version: 1.0.0
  contact:
    name: Redis Streams POC
    url: https://github.com/your-username/redis-streams-poc
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-domain.com
    description: Production server
...
paths:
...
  /api/monitoring/streams:
    get:
      tags:
        - Monitoring
      summary: Get all Redis Streams information
      description: Get detailed information about all Redis Streams including length, consumer groups, and recent messages
      responses:
        '200':
          description: Streams information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  streams:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        length:
                          type: integer
                          example: 5
                        info:
                          type: object
                          description: Stream metadata
                        groups:
                          type: array
                          items:
                            type: object
                        recentMessages:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                                example: "1758619840865-0"
                              data:
                                type: object
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/monitoring/streams/{streamName}:
    get:
      tags:
        - Monitoring
      summary: Get specific stream information
      description: Get detailed information about a specific Redis Stream
      parameters:
        - name: streamName
          in: path
          required: true
          description: Name of the Redis Stream
          schema:
            type: string
            example: user:events
        - name: limit
          in: query
          description: Maximum number of recent messages to return
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Stream information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  stream:
                    type: string
                    example: user:events
                  length:
                    type: integer
                    example: 5
                  info:
                    type: object
                    description: Stream metadata
                  groups:
                    type: array
                    items:
                      type: object
                  recentMessages:
                    type: array
                    items:
                      type: object
                  messageRange:
                    type: array
                    items:
                      type: object
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/monitoring/streams/{streamName}/groups:
    get:
      tags:
        - Monitoring
      summary: Get consumer groups for a stream
      description: Get information about consumer groups for a specific Redis Stream
      parameters:
        - name: streamName
          in: path
          required: true
          description: Name of the Redis Stream
          schema:
            type: string
            example: user:events
      responses:
        '200':
          description: Consumer groups information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  stream:
                    type: string
                    example: user:events
                  groups:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          example: user-processing-group
                        consumers:
                          type: integer
                          example: 2
                        pending:
                          type: integer
                          example: 0
                        consumers:
                          type: array
                          items:
                            type: object
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/monitoring/streams/{streamName}/messages:
    get:
      tags:
        - Monitoring
      summary: Read messages from stream
      description: Read messages from a specific Redis Stream
      parameters:
        - name: streamName
          in: path
          required: true
          description: Name of the Redis Stream
          schema:
            type: string
            example: user:events
        - name: start
          in: query
          description: Starting message ID
          required: false
          schema:
            type: string
            default: "0"
        - name: end
          in: query
          description: Ending message ID
          required: false
          schema:
            type: string
            default: "+"
        - name: count
          in: query
          description: Maximum number of messages to return
          required: false
          schema:
            type: integer
            default: 10
        - name: direction
          in: query
          description: Direction to read messages
          required: false
          schema:
            type: string
            enum: [forward, backward]
            default: forward
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  stream:
                    type: string
                    example: user:events
                  direction:
                    type: string
                    example: forward
                  count:
                    type: integer
                    example: 5
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "1758619840865-0"
                        data:
                          type: object
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/monitoring/redis/info:
    get:
      tags:
        - Monitoring
      summary: Get Redis server information
      description: Get information about the Redis server including version, memory usage, and statistics
      responses:
        '200':
          description: Redis server information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  redis:
                    type: object
                    properties:
                      version:
                        type: string
                        example: "7.0.0"
                      mode:
                        type: string
                        example: "standalone"
                      uptime:
                        type: string
                        example: "3600"
                      connectedClients:
                        type: string
                        example: "5"
                      usedMemory:
                        type: string
                        example: "1.2M"
                      totalCommands:
                        type: string
                        example: "1000"
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: User not found
        details:
          type: string
          example: The requested user ID does not exist
        path:
          type: string
          example: /api/users/123e4567-e89b-12d3-a456-426614174000
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...